Break through deployment mode
When deployment mode is detected we patch Bundler to act as if it
wasn't set.

This:
- makes Bundler not set `use_system_gems?` to `false`
- makes Bundler not set the vendored path to `vendor/bundle`

But thanks to the `GEM_HOME` and `GEM_PATH` variables that have been set
beforehand, gems will be looked up in the appropriate location, all the
while having broken out of being able to see _only_ vendored gems.

Ergo, gem injection can proceed.