Enforce RUBYOPT order
Since Ruby 1.9 `rubygems` is automatically required default. Ruby and
`rubygems` contain mechanisms that enforce rubygems is not just loaded,
but that it is being loaded *first*.

Right before executing the Ruby process we make sure that the Ruby
injector is appearing first, and whatever else is present still appears
too, but afterwards.