Enable injector second stage
When `bundle exec` happens, it proceeds with processing Bundler things,
then `exec`s into the actual Ruby program to execute.

When it does so, this is a new process, hence any patch that we have
applied goes away, which results in deployment mode being re-armed.

Apply the `bundler` patches to override Bundler behaviours.

We can safely load `bundler` withotu forking in that case since we are
in a bundled case.